+++ docs/conf.py
@@ -12,9 +12,6 @@
 #
 import os
 import sys
-from datetime import date
-
-from colordetect import ColorDetect
 
 file_loc = os.path.split(__file__)[0]
 sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(file_loc), ".")))
@@ -27,7 +24,7 @@ copyright = "2020, Marvin Kweyu"
 author = "Marvin Kweyu"
 
 # The full version, including alpha/beta/rc tags
-release = "1.2.0"
+release = "1.3.0rc"
 
 # -- General configuration ---------------------------------------------------